backend/requirements.txt
pandas
requests
python-dotenv
pandas_datareader